refactor: moved executable scripts out of src dir
