+++ apps/native-mcp/package.json
@@ -39,7 +39,6 @@
     "build:api": "wrangler deploy --dry-run --outdir dist-api",
     "dev:api": "dotenv -- wrangler dev src/api.ts --env development --port 8788",
     "dev:stdio": "NODE_ENV=development tsx --watch src/mcp/stdio.ts",
-    "dev:mastra": "mastra dev --dir ./mastra",
     "dev": "npm run dev:api",
     "test": "vitest run",
     "test:watch": "vitest",
@@ -56,8 +55,8 @@
     "format": "prettier --write \"src/**/*.{ts,tsx,json}\"",
     "mcp:stdio": "NODE_ENV=development tsx src/mcp/stdio.ts",
     "inspect": "NODE_ENV=development CLIENT_PORT=6275 SERVER_PORT=6278 npx @modelcontextprotocol/inspector -e HEROUI_NATIVE_API_URL=http://localhost:8788 tsx src/mcp/stdio.ts",
-    "extract:components": "tsx src/extraction/cli/extract-components.ts",
-    "extract:theme": "tsx src/extraction/cli/extract-theme.ts",
+    "extract:components": "tsx scripts/extract-components.ts",
+    "extract:theme": "tsx scripts/extract-theme.ts",
     "extract:dev": "./scripts/extract-dev.sh",
     "extract:dev:components": "./scripts/extract-dev.sh components",
     "extract:dev:theme": "./scripts/extract-dev.sh theme",
