+++ apps/native-mcp/tsconfig.json
@@ -7,6 +7,6 @@
       "@shared/*": ["./src/shared/*"]
     }
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx", "lib/**/*.ts", "vitest.config.ts"],
+  "include": ["src/**/*.ts", "src/**/*.tsx", "scripts/**/*.ts", "lib/**/*.ts", "vitest.config.ts"],
   "exclude": ["node_modules", "dist", "mastra", ".mastra"]
 }
