+++ apps/react-mcp/.env.example
@@ -1,5 +1,5 @@
 # API service URL
-HEROUI_NATIVE_API_URL=http://localhost:8787
+HEROUI_API_URL=http://localhost:8787
 
 # Cloudflare R2 Storage Configuration
 # Required for both extraction and API deployment
