+++ apps/react-mcp/package.json
@@ -53,11 +53,11 @@
     "format": "prettier --write \"src/**/*.{ts,tsx,json}\"",
     "mcp:stdio": "NODE_ENV=development tsx src/mcp/stdio.ts",
     "inspect": "NODE_ENV=development npx @modelcontextprotocol/inspector -e HEROUI_API_URL=http://localhost:8787 tsx src/mcp/stdio.ts",
-    "extract:components": "tsx src/extraction/cli/extract-components.ts",
-    "extract:theme": "tsx src/extraction/cli/extract-theme.ts",
-    "extract:dev": "./src/extraction/cli/extract-dev.sh",
-    "extract:dev:components": "./src/extraction/cli/extract-dev.sh components",
-    "extract:dev:theme": "./src/extraction/cli/extract-dev.sh theme",
+    "extract:components": "tsx scripts/extract-components.ts",
+    "extract:theme": "tsx scripts/extract-theme.ts",
+    "extract:dev": "./scripts/extract-dev.sh",
+    "extract:dev:components": "./scripts/extract-dev.sh components",
+    "extract:dev:theme": "./scripts/extract-dev.sh theme",
     "clean": "rm -rf dist",
     "prepublishOnly": "pnpm build",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s --commit-path .",
