+++ apps/react-mcp/scripts/extract-dev.sh
